gast löschen

+++ project-barney/src/dashboard/projekte_ansicht_orga.php_modals/gast_loeschen.php
@@ -8,12 +8,8 @@ try {
     $dbh = new PDO($dsn, $user, $password);
     $dbh->setAttribute( PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION );
     //Testen ob es den nutzer schon gibt
-    $Stmt = $dbh->prepare("DELETE FROM produktliste WHERE projektid = ?");
-    $Stmt->execute([intval($_POST['projektid'])]);
-    $Stmt = $dbh->prepare("DELETE FROM projektuser WHERE projektid = ?");
-    $Stmt->execute([intval($_POST['projektid'])]);
-    $Stmt = $dbh->prepare("DELETE FROM projekte WHERE projektid = ?");
-    $Stmt->execute([intval($_POST['projektid'])]);
+    $Stmt = $dbh->prepare("DELETE FROM projektuser WHERE projektid = ? AND userid = ?");
+    $Stmt->execute([intval($_POST['projektid']), [intval($_POST['userid'])]);
 
     header("Location: ../projekte.php");
     die();
